Add files via upload
All the packages have been uploaded in compressed format due to their large sizes. Rosflight and rosflight_plugins are the packages needed to run all the dependencies associated with the Naze32 firmware. A more complete description of the same can be found at https://github.com/rosflight/rosflight and http://rosflight.org/ 

Rosplane_sim contains the Gazebo launch file which, in turn, launches the following:

1) Gazebo world and fixed wing model 
2) A series of Python nodes developed by us for the purpose of this project. These files are - param.py, GPS_sim.py, autopilot4.py, path_follow2.py, state_estimation.py. 

In addition to in-built ROS messages, we have defined a set of custom messages within the 'rosplane_msgs' folder that helps in communicating various topics between the aforementioned nodes.